chore: move charm package to application/charm
