+++ domain/application/service/manifest.go
@@ -6,8 +6,8 @@ package service
 import (
 	"fmt"
 
+	"github.com/juju/juju/domain/application/charm"
 	applicationerrors "github.com/juju/juju/domain/application/errors"
-	"github.com/juju/juju/domain/charm"
 	internalcharm "github.com/juju/juju/internal/charm"
 )
