+++ domain/application/state/actions.go
@@ -5,7 +5,7 @@ package state
 
 import (
 	corecharm "github.com/juju/juju/core/charm"
-	"github.com/juju/juju/domain/charm"
+	"github.com/juju/juju/domain/application/charm"
 )
 
 func decodeActions(actions []charmAction) charm.Actions {
