+++ domain/application/state/actions_test.go
@@ -6,7 +6,7 @@ package state
 import (
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/domain/charm"
+	"github.com/juju/juju/domain/application/charm"
 	schematesting "github.com/juju/juju/domain/schema/testing"
 )
