+++ domain/application/state/metadata.go
@@ -9,7 +9,7 @@ import (
 	"github.com/juju/version/v2"
 
 	corecharm "github.com/juju/juju/core/charm"
-	"github.com/juju/juju/domain/charm"
+	"github.com/juju/juju/domain/application/charm"
 )
 
 type relationKind = string
