+++ domain/application/types.go
